feat: 添加单例client，分离出handler

+++ user_service/pojo/user.go
@@ -1,7 +1,7 @@
 package pojo
 
 type User struct {
-	Id        int    `gorm:"primaryKey"`
+	Id        int64  `gorm:"primaryKey"`
 	Name      string `gorm:"column:name"`
 	Avatar    string `gorm:"column:avatar"`
 	Signature string `gorm:"column:signature"`
